# docs/api/topic-e-plan-2-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinBrain OS Plan 2 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Agents",
        "operationId": "list_agents_agents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/agents/runs": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Start Run",
        "operationId": "start_run_agents_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/agents/runs/{run_id}/events": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Run Events",
        "operationId": "run_events_agents_runs__run_id__events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events; each data line is one AgentRunEvent",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string",
                      "contentMediaType": "application/json",
                      "contentSchema": {
                        "$ref": "#/components/schemas/AgentRunEvent"
                      }
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{agent_id}/autonomy": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Change Autonomy",
        "operationId": "change_autonomy_agents__agent_id__autonomy_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutonomyChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentCardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/kill-switch": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Kill Switch",
        "operationId": "kill_switch_agents_kill_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/agents/journey": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Journey",
        "operationId": "journey_agents_journey_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AgentCard": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "job_function": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobFunction"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewer_job_function": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobFunction"
              },
              {
                "type": "null"
              }
            ]
          },
          "build_status": {
            "type": "string",
            "enum": [
              "built",
              "designed"
            ],
            "title": "Build Status"
          },
          "autonomy_level": {
            "$ref": "#/components/schemas/AutonomyLevel"
          },
          "scoped_autonomy": {
            "items": {
              "$ref": "#/components/schemas/ScopedAutonomy"
            },
            "type": "array",
            "title": "Scoped Autonomy"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/AgentSkill"
            },
            "type": "array",
            "title": "Skills"
          },
          "kill_switch_engaged": {
            "type": "boolean",
            "title": "Kill Switch Engaged"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentMetrics"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "purpose",
          "job_function",
          "reviewer_job_function",
          "build_status",
          "autonomy_level",
          "scoped_autonomy",
          "skills",
          "kill_switch_engaged",
          "metrics"
        ],
        "title": "AgentCard"
      },
      "AgentCardResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "agent": {
            "$ref": "#/components/schemas/AgentCard"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "agent"
        ],
        "title": "AgentCardResponse"
      },
      "AgentListResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "global_kill_switch_engaged": {
            "type": "boolean",
            "title": "Global Kill Switch Engaged"
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentCard"
            },
            "type": "array",
            "title": "Agents"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "global_kill_switch_engaged",
          "agents"
        ],
        "title": "AgentListResponse"
      },
      "AgentMetrics": {
        "properties": {
          "proposals": {
            "type": "integer",
            "title": "Proposals"
          },
          "approved_unedited": {
            "type": "integer",
            "title": "Approved Unedited"
          },
          "approved_edited": {
            "type": "integer",
            "title": "Approved Edited"
          },
          "rejected": {
            "type": "integer",
            "title": "Rejected"
          },
          "unedited_approval_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unedited Approval Rate"
          },
          "promotion_recommended": {
            "type": "boolean",
            "title": "Promotion Recommended"
          }
        },
        "type": "object",
        "required": [
          "proposals",
          "approved_unedited",
          "approved_edited",
          "rejected",
          "unedited_approval_rate",
          "promotion_recommended"
        ],
        "title": "AgentMetrics"
      },
      "AgentRunCreated": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "events_url": {
            "type": "string",
            "title": "Events Url"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "run_id",
          "events_url"
        ],
        "title": "AgentRunCreated"
      },
      "AgentRunEvent": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "type": {
            "type": "string",
            "enum": [
              "run_started",
              "tool_called",
              "proposal_created",
              "waiting_for_review",
              "run_completed"
            ],
            "title": "Type"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "action_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Id"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "sequence",
          "type",
          "agent_id",
          "message"
        ],
        "title": "AgentRunEvent"
      },
      "AgentRunRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Goal"
          }
        },
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "AgentRunRequest"
      },
      "AgentSkill": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "wave": {
            "type": "string",
            "enum": [
              "W1",
              "W2"
            ],
            "title": "Wave"
          },
          "availability": {
            "type": "string",
            "enum": [
              "available",
              "planned"
            ],
            "title": "Availability"
          },
          "side_effect": {
            "type": "string",
            "enum": [
              "read",
              "draft",
              "external",
              "money"
            ],
            "title": "Side Effect"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "wave",
          "availability",
          "side_effect"
        ],
        "title": "AgentSkill"
      },
...
      "AutonomyChangeRequest": {
        "properties": {
          "action": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z_]+$",
            "title": "Action"
          },
          "level": {
            "$ref": "#/components/schemas/AutonomyLevel"
          },
          "max_amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Amount"
          }
        },
        "type": "object",
        "required": [
          "action",
          "level"
        ],
        "title": "AutonomyChangeRequest"
      },
      "AutonomyLevel": {
        "type": "string",
        "enum": [
          "L0",
          "L1",
          "L2",
          "L3"
        ],
        "title": "AutonomyLevel"
      },
...
      "DataMode": {
        "type": "string",
        "enum": [
          "stub",
          "live"
        ],
        "title": "DataMode",
        "description": "Every contract response says whether its data is a canned stub or computed live."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobFunction": {
        "type": "string",
        "enum": [
          "owner",
          "operations",
          "finance",
          "sales",
          "customer_service",
          "marketing",
          "procurement",
          "logistics",
          "production",
          "hr",
          "compliance"
        ],
        "title": "JobFunction",
        "description": "The positions a person can hold. One person may hold several."
      },
      "JourneyAgent": {
        "properties": {
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "autonomy_level": {
            "$ref": "#/components/schemas/AutonomyLevel"
          },
          "build_status": {
            "type": "string",
            "enum": [
              "built",
              "designed"
            ],
            "title": "Build Status"
          },
          "override_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Rate"
          },
          "estimated_hours_saved": {
            "type": "number",
            "title": "Estimated Hours Saved"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "name",
          "autonomy_level",
          "build_status",
          "override_rate",
          "estimated_hours_saved"
        ],
        "title": "JourneyAgent"
      },
      "JourneyFunction": {
        "properties": {
          "job_function": {
            "$ref": "#/components/schemas/JobFunction"
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/JourneyAgent"
            },
            "type": "array",
            "title": "Agents"
          }
        },
        "type": "object",
        "required": [
          "job_function",
          "agents"
        ],
        "title": "JourneyFunction"
      },
      "JourneyResponse": {
        "properties": {
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "estimate_note": {
            "type": "string",
            "title": "Estimate Note"
          },
          "functions": {
            "items": {
              "$ref": "#/components/schemas/JourneyFunction"
            },
            "type": "array",
            "title": "Functions"
          }
        },
        "type": "object",
        "required": [
          "data_mode",
          "estimate_note",
          "functions"
        ],
        "title": "JourneyResponse"
      },
      "KillSwitchRequest": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id",
            "description": "None means every agent"
          },
          "engaged": {
            "type": "boolean",
            "title": "Engaged"
          }
        },
        "type": "object",
        "required": [
          "engaged"
        ],
        "title": "KillSwitchRequest"
      },
...
      "ScopedAutonomy": {
        "properties": {
          "action": {
            "type": "string",
            "title": "Action"
          },
          "level": {
            "$ref": "#/components/schemas/AutonomyLevel"
          },
          "max_amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Amount"
          }
        },
        "type": "object",
        "required": [
          "action",
          "level"
        ],
        "title": "ScopedAutonomy"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}